+++ src/main/java/com/iexec/core/chain/DealWatcherService.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2020-2023 IEXEC BLOCKCHAIN TECH
+ * Copyright 2020-2024 IEXEC BLOCKCHAIN TECH
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -188,7 +188,7 @@ public class DealWatcherService {
      */
     private void handleDeal(DealEvent dealEvent) {
         String chainDealId = dealEvent.getChainDealId();
-        Optional<ChainDeal> oChainDeal = iexecHubService.getChainDeal(chainDealId);
+        Optional<ChainDeal> oChainDeal = iexecHubService.getChainDealWithDetails(chainDealId);
         if (oChainDeal.isEmpty()) {
             log.error("Could not get chain deal [chainDealId:{}]", chainDealId);
             return;